Delete Templates directory
The Templates direction is created by the code but should not be uploaded to the GitHub repo because of quota restriction and issues with big files. It should remain a local folder and only the template zip file should be added as asset to releases (but the to the repo).